refactor: split admin_commands.py monolith, remove dead code, migrate config to pydantic
Phase A — Dead code removal:
- Delete bot/application/slots_service.py (232 lines, unreferenced)
- Delete bot/application/slots_custom_functions.py (85 lines, unreferenced)

Phase B — Split 1335-line admin_commands.py into focused modules:
- handlers/_admin_utils.py  — shared helpers and _unmute_user
- handlers/mute.py          — /mute, /amute, /selfmute, /unmute
- handlers/tag.py           — /tag
- handlers/transfer.py      — /transfer
- handlers/protect.py       — /protect, /unprotect, cb_protect
- handlers/admin_score.py   — /reset, /set, /add, /sub
- handlers/admin_user.py    — /save, /restore, /op, /deop
- handlers/help.py          — /help, cb_help
Also fixes typo: allback_data → callback_data in protect cancel button

Phase E — Config loading via pydantic model_validate:
- Replace all @dataclass config classes with pydantic BaseModel subclasses
- Add @field_validator on AdminConfig.users for username normalization
- Reduce load_config() to a single AppConfig.model_validate(raw) call
